mpharma/requirements.txt
Django==1.11.13
django-filter==1.1.0
djangorestframework==3.6.4
djangorestframework-filters==0.10.2
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
redis==2.10.3
celery==3.1.18
